Add wildcards to composite action constructors
Composite action list-constructors were previously implemented with
a param of type List<Action>. This meant that a list containing a
subtype of Action was ineligible for use. Now, an upper-bounded wildcard
is used instead, which allows a list of any subtype of Action to be
used. To implement this, and because it's good general practice, the
List is now (shallowly) copied to an ArrayList, not merely reused.