Add wildcards to constructors (#3)
List-constructors were previously implemented with a param of type List<X>. This meant that a list containing a subtype of X was ineligible for use. Now, an upper-bounded wildcard is used instead, which allows a list of any subtype of X to be used. To implement this, and because it's good general practice, the List is now (shallowly) copied to an ArrayList, not merely reused.