Add LGPL license per discussion
The GPL was deemed too restrictive because it requires all re-users to
distribute all their linked code under the same license. Several
programers objected to the possibility of their code being put into a
proprietary project with closed-source modifications, or even sold,
as might occur under an MIT or BSD license. The LGPL was adopted as a
compromise, because it requires modifications to be made public while
still allowing inclusion in projects with other license schemes.