Action tests (#4)
This commit performs two separate but related changes:

1. It adds tests for all concrete actions. It tests that RunnerAction
runs its runnable, that ParallelAction runs its actions at the same
time (using a CountDownLatch), and that SerialAction runs its actions
in the correct order.

2. The abstract action base classes were in the auto package, not the
auto.actions package. There was no clear justification for this layout,
and it lead to unnecessary imports. This commit moves those classes so
that the entire autonomous action library is self-contained.